PluginComponent在重新加载模板时未执行aboutToDisappear

+++ frameworks/core/components_ng/pattern/custom/custom_node_base.cpp
@@ -29,7 +29,7 @@ CustomNodeBase::~CustomNodeBase()
     if (!executeFireOnAppear_ && appearFunc_) {
         appearFunc_();
     }
-    if (destroyFunc_) {
+    if (!executeFireOnAppear_ && destroyFunc_) {
         destroyFunc_();
     }
 }
